moved BulkUploadResponseWrapper to models-- the bulk (haha) of the logic is unchanged, just adds a check for problem rows and a todo for later
