+++ corehq/apps/users/views/__init__.py
@@ -1,5 +1,4 @@
 import json
-from collections import defaultdict
 from datetime import datetime
 
 import langcodes
@@ -113,6 +112,7 @@ from corehq.apps.users.models import (
     Invitation,
     UserRole,
     WebUser,
+    BulkUploadResponseWrapper,
 )
 from corehq.apps.users.tasks import (
     bulk_download_users_async,
@@ -1265,31 +1265,7 @@ def web_user_upload_job_poll(request, domain, download_id,
         'user_type': _('web users'),
     })
 
-    class _BulkUploadResponseWrapper(object):
-
-        def __init__(self, context):
-            results = context.get('result') or defaultdict(lambda: [])
-            self.response_rows = results['rows']
-            self.response_errors = results['errors']
-            self.problem_rows = [r for r in self.response_rows if r['flag'] not in ('updated', 'invited')]
-
-        def success_count(self):
-            return len(self.response_rows) - len(self.problem_rows)
-
-        def has_errors(self):
-            return bool(self.response_errors or self.problem_rows)
-
-        def errors(self):
-            errors = []
-            for row in self.problem_rows:
-                if row['flag'] == 'missing-data':
-                    errors.append(_('A row with no username was skipped'))
-                else:
-                    errors.append('{username}: {flag}'.format(**row))
-            errors.extend(self.response_errors)
-            return errors
-
-    context['result'] = _BulkUploadResponseWrapper(context)
+    context['result'] = BulkUploadResponseWrapper(context)
     return render(request, template, context)
